Update the hostname truncation example, also MUST truncate on character boundary
Still struggling with SVG, I thought I had conqueued it but
it thought the "o" was cute, so come up with a different example.

Add text about truncating on a character boundary and give an
example about omitting the first octet of a 2-octet UTF-8 character
and padding with NUL instead.

Fixes #26